requirements.txt
sqlalchemy
pymysql
streamlit_autorefresh
difflib